♻️ refactor(server): collapse JWT to a single commonMain file; move :contract dep to commonMain
